% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGP_plots.R
\name{AGP_plotAGP}
\alias{AGP_plotAGP}
\title{Plot Ambulatory Glucose Profile (AGP)}
\usage{
AGP_plotAGP(
  data,
  ID,
  time,
  glucose,
  target_glucose = c(70, 180),
  smooth = TRUE,
  span = 0.3,
  title = FALSE
)
}
\arguments{
\item{data}{A data frame containing CGM readings.}

\item{ID}{A character string specifying the column name representing the subject ID.}

\item{time}{A character string specifying the column name representing the timestamp of CGM readings.}

\item{glucose}{A character string specifying the column name representing glucose values.}

\item{target_glucose}{A numeric vector of length 2 specifying the lower and upper bounds for the target glucose range (default: `c(70, 180)`).}

\item{smooth}{A logical indicating whether to smooth the glucose percentiles using LOESS (default: `TRUE`).}

\item{span}{A numeric value controlling the smoothness of LOESS fitting (default: `0.3`).}

\item{title}{A logical indicating whether to include the subject ID as the title (default: `FALSE`).}
}
\value{
A ggplot2 object representing an AGP plot.
}
\description{
This function generates an Ambulatory Glucose Profile (AGP) plot summarizing
glucose fluctuations over a 24-hour period. The plot includes the median, interquartile
range, and variability percentiles (5th-95th). It integrates CGMS2DayByDay functionality.
}
