% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estTIR.R
\name{naive_est}
\alias{naive_est}
\title{Naive Estimator for Time in Range (TIR)}
\usage{
naive_est(
  data,
  min_time = 0,
  max_time = (1440 * 7 - 5),
  boot = NULL,
  id_col = "patient_id",
  time = "time",
  value_in_range = "value_in_range"
)
}
\arguments{
\item{data}{A data frame containing glucose readings.}

\item{min_time}{Minimum time value to filter data.}

\item{max_time}{Maximum time value to filter data.}

\item{boot}{Number of bootstrap replicates (default: NULL).}

\item{id_col}{Column name for subject ID.}

\item{time}{Column name for time variable.}

\item{value_in_range}{Column name for 0/1 indicator of glucose within range.}
}
\value{
A list with estimated TIR, optional bootstrap standard error and confidence intervals.
}
\description{
Computes the mean Time in Range (TIR) across subjects using a naive estimator. Optionally supports bootstrapping.
}
