% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estTIR.R
\name{proposed_est_cox}
\alias{proposed_est_cox}
\title{Proposed Estimator Using Cox Model}
\usage{
proposed_est_cox(
  data,
  min_time = 0,
  max_time = (1440 * 7 - 5),
  id_col = "patient_id",
  event_col = "event",
  start_col = "minute_enrollment",
  stop_col = "time2",
  formula = "var1",
  boot = NULL,
  value_in_range = "value_in_range",
  period = 5
)
}
\arguments{
\item{data}{A data frame of glucose and survival information.}

\item{min_time}{Minimum time value for filtering.}

\item{max_time}{Maximum time value for filtering.}

\item{id_col}{Subject ID column.}

\item{event_col}{Event column for Cox model.}

\item{start_col}{Start time column.}

\item{stop_col}{Stop time column.}

\item{formula}{Right-hand side formula for Cox model.}

\item{boot}{Number of bootstrap iterations.}

\item{value_in_range}{Column name for binary glucose range indicator.}
}
\value{
A list with TIR estimate, standard error, and confidence intervals.
}
\description{
Estimates Time in Range using survival-weighted averaging based on a Cox model for dropout. Supports bootstrapping.
}
