% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estTIR.R
\name{proposed_est_noninfo}
\alias{proposed_est_noninfo}
\title{Proposed Estimator (Non-informative Follow-up)}
\usage{
proposed_est_noninfo(
  data,
  min_time = 0,
  max_time = (1440 * 7 - 5),
  boot = NULL,
  id_col = "patient_id",
  time = "time",
  value_in_range = "value_in_range"
)
}
\arguments{
\item{data}{A data frame of glucose time-series.}

\item{min_time}{Minimum time threshold.}

\item{max_time}{Maximum time threshold.}

\item{boot}{Number of bootstrap replicates.}

\item{id_col}{Column for subject ID.}

\item{time}{Time column.}

\item{value_in_range}{Column name for 0/1 indicator for in-range glucose.}
}
\value{
A list containing TIR estimate, standard error, and confidence intervals.
}
\description{
Computes the mean TIR using time-stratified averaging. Supports optional bootstrapping.
}
