% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_CGM_data.R
\name{read_CGM_data}
\alias{read_CGM_data}
\title{Read and Process CGM Data}
\usage{
read_CGM_data(file, ID, time, glucose, time_interval = c(1440, 5760))
}
\arguments{
\item{file}{Character. Path to the CSV file.}

\item{ID}{Character. Column name representing patient ID.}

\item{time}{Character. Column name representing date-time values.}

\item{glucose}{Character. Column name representing glucose readings.}

\item{time_interval}{Numeric vector of length 2. Time range in minutes for filtering data (default: c(1440, 5760)).}
}
\value{
A cleaned data frame with processed glucose readings.
}
\description{
This function reads a CGM dataset, processes the time column, computes enrollment time in minutes and days,
rounds time intervals, removes duplicate and non-numeric glucose values, and filters data within a specified time interval.
}
