% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estTIR.R
\name{value_in_range_vec}
\alias{value_in_range_vec}
\title{Vectorized Value-in-Range Check}
\usage{
value_in_range_vec(x, range = c(-.Machine$integer.max, .Machine$integer.max))
}
\arguments{
\item{x}{A numeric vector.}

\item{range}{A numeric vector of length 2 indicating the lower and upper bounds.}
}
\value{
An integer vector: 1 if in range, 0 if not, NA if input is NA.
}
\description{
Checks whether a numeric vector falls within a specified range.
}
